ideas

+++ src/idea/idea.service.ts
@@ -1,5 +1,5 @@
 import { IdeaEntity } from './idea.entity';
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import {Repository} from 'typeorm'
 import { InjectRepository } from '@nestjs/typeorm';
 import { IdeaDTO } from './idea.dto';
@@ -19,17 +19,29 @@ export class IdeaService {
     }
 
     async read(id:string){
-        return await this.ideaRepository.findOne({where:{id}}); 
+        const idea= await this.ideaRepository.findOne({where:{id}}); 
+        if(!idea){
+            throw new HttpException("Not found",HttpStatus.NOT_FOUND);
+        }
+        return idea;
     }
 
     async update(id:string,data:Partial<IdeaDTO>){
+        const idea= await this.ideaRepository.findOne({where:{id}}); 
+        if(!idea){
+            throw new HttpException("Not found",HttpStatus.NOT_FOUND);
+        }
         await this.ideaRepository.update({id},data);
-        return await this.ideaRepository.findOne({id});
+        return idea;
     }
 
     async destroy(id:string){
+        const idea= await this.ideaRepository.findOne({where:{id}}); 
+        if(!idea){
+            throw new HttpException("Not found",HttpStatus.NOT_FOUND);
+        }
         await this.ideaRepository.delete({id});
-        return {deleted:true};
+        return idea;
     }
 }
